refactor(learn): split pattern-cache.ts into ranking / cache-store / retrieval (#3721)
pattern-cache.ts co-located five concerns behind a wide interface; testing a
pure helper like rankPatterns or perfWeight required mocking three modules
(db/internal, settings, logger) just to import them, because the file mixed
db/settings/logger imports with the pure math.

Split into focused modules:
  - pattern-ranking.ts   — PURE: extractKeywords, scorePattern, perfWeight,
                           rankPatterns. Zero db/settings/logger imports (only a
                           type-only ApprovedPatternRow), so its test imports it
                           directly with NO mock.module() at all.
  - pattern-cache-store.ts — the TTL/LRU store + getCachedPatterns +
                           invalidatePatternCache; changing eviction policy now
                           touches no ranking code.
  - pattern-cache.ts     — slim retrieval composition (buildRetrievalQuery,
                           getRelevantPatterns, buildLearnedPatternsSection) +
                           public barrel re-exporting the prior surface, so
                           agent.ts / admin-learned-patterns.ts /
                           org-knowledge-section.ts and the existing test mocks
                           import from pattern-cache unchanged.

pattern-ranking.test.ts rewritten to import the pure module with no mocks (the
removal is the testability win). The trivial DEFAULT_LATENCY_BUDGET_MS>0
assertion dropped — that default lives in learn-settings and is covered there.
No behavior change.

Deferred from v0.0.17 "Performance-aware Atlas" (Architecture Backlog).
