+++ packages/api/src/lib/learn/pattern-cache.ts
@@ -1,12 +1,20 @@
 /**
- * In-memory cache for approved learned patterns with TTL-based expiry
- * and keyword-based relevance filtering.
+ * Learned-pattern retrieval: compose the cache store, the pure ranking math, and
+ * the settings resolver into the agent-facing API + public barrel.
  *
- * Cache is keyed by orgId (null for global). Invalidated by admin
- * approve/reject actions via `invalidatePatternCache()`.
+ * The concerns this file used to co-locate now live in focused modules (#3721):
+ *   - cache store (TTL/LRU + invalidation) → `pattern-cache-store.ts`
+ *   - keyword + perf-weighted ranking (pure) → `pattern-ranking.ts`
+ *   - ATLAS_LEARN_* settings reads → `learn-settings.ts` (#3722)
+ *   - pattern → prompt rendering → `pattern-format.ts` (#3720)
+ *
+ * This module owns only the retrieval-query assembly and the two retrieval
+ * entry points (`getRelevantPatterns`, `buildLearnedPatternsSection`). It also
+ * re-exports the prior public surface so consumers (agent.ts,
+ * admin-learned-patterns.ts, org-knowledge-section.ts) and existing test mocks
+ * keep importing from `pattern-cache` unchanged.
  */
 
-import { getApprovedPatterns, type ApprovedPatternRow } from "@atlas/api/lib/db/internal";
 import { createLogger } from "@atlas/api/lib/logger";
 import { renderPattern } from "./pattern-format";
 import {
@@ -16,135 +24,28 @@ import {
   getConfidenceThreshold,
   getLatencyBudgetMs,
 } from "./learn-settings";
+import { extractKeywords, perfWeight, rankPatterns, type ScoredPattern } from "./pattern-ranking";
+import { getCachedPatterns, invalidatePatternCache, _resetPatternCache } from "./pattern-cache-store";
 
-// All ATLAS_LEARN_* reads now live in the single learn-settings resolver
-// (#3722). Re-export the retrieval knobs from this module's public surface so
-// existing consumers (agent.ts) and the test mocks that import them from
-// pattern-cache keep working unchanged.
+// Public barrel — these moved into focused modules during the #3721 split (and
+// #3722 for the settings reads), but agent.ts, admin-learned-patterns.ts,
+// org-knowledge-section.ts, and a number of test mocks import them from
+// `pattern-cache`, so the names stay stable on this module's surface.
 export {
   DEFAULT_RETRIEVAL_TURNS,
   DEFAULT_LATENCY_BUDGET_MS,
   getRetrievalTurns,
   getConfidenceThreshold,
   getLatencyBudgetMs,
+  extractKeywords,
+  perfWeight,
+  rankPatterns,
+  invalidatePatternCache,
+  _resetPatternCache,
 };
+export type { ScoredPattern };
 
 const log = createLogger("pattern-cache");
-
-// ---------------------------------------------------------------------------
-// Cache
-// ---------------------------------------------------------------------------
-
-const DEFAULT_TTL_MS = 5 * 60 * 1000; // 5 minutes
-const MAX_ENTRIES = 500;
-
-interface CacheEntry {
-  patterns: ApprovedPatternRow[];
-  expiresAt: number;
-  lastAccessedAt: number;
-}
-
-const cache = new Map<string, CacheEntry>();
-
-/** Org segment of a cache key — `"__global__"` for null orgId, prefixed to
- *  avoid collision. */
-function orgKeyPart(orgId: string | null): string {
-  return orgId === null ? "__global__" : `org:${orgId}`;
-}
-
-/** Canonical cache key — scoped by org AND connection group (#3611) so a
- *  `us-prod` agent session never serves a `eu-prod` group's cached patterns.
- *  `"__nogroup__"` represents the default (flat `entities/`) scope. */
-function cacheKey(orgId: string | null, connectionGroupId: string | null): string {
-  const groupPart = connectionGroupId === null ? "__nogroup__" : `group:${connectionGroupId}`;
-  return `${orgKeyPart(orgId)}::${groupPart}`;
-}
-
-/** Get approved patterns for an org + connection group, hitting cache first.
- *  DB failures are logged and return [] without being cached. */
-async function getCachedPatterns(
-  orgId: string | null,
-  connectionGroupId: string | null,
-): Promise<ApprovedPatternRow[]> {
-  const key = cacheKey(orgId, connectionGroupId);
-  const entry = cache.get(key);
-
-  if (entry && Date.now() < entry.expiresAt) {
-    entry.lastAccessedAt = Date.now();
-    return entry.patterns;
-  }
-
-  try {
-    const patterns = await getApprovedPatterns(orgId, connectionGroupId);
-
-    // Evict oldest entry if cache is at capacity
-    if (cache.size >= MAX_ENTRIES) {
-      let oldestKey: string | undefined;
-      let oldestTime = Infinity;
-      for (const [k, v] of cache) {
-        if (v.lastAccessedAt < oldestTime) {
-          oldestTime = v.lastAccessedAt;
-          oldestKey = k;
-        }
-      }
-      if (oldestKey) cache.delete(oldestKey);
-    }
-
-    const now = Date.now();
-    cache.set(key, { patterns, expiresAt: now + DEFAULT_TTL_MS, lastAccessedAt: now });
-    return patterns;
-  } catch (err) {
-    log.warn(
-      { orgId, connectionGroupId, err: err instanceof Error ? err.message : String(err) },
-      "Failed to load approved patterns (table may not exist yet) — not caching",
-    );
-    return [];
-  }
-}
-
-/**
- * Invalidate the pattern cache for a specific org, across ALL of its connection
- * groups. The admin approve/reject path (`admin-learned-patterns.ts`) operates
- * at org granularity and doesn't know which group a reviewed pattern belongs
- * to, so a single call must clear every group-scoped entry for the org (#3611).
- */
-export function invalidatePatternCache(orgId: string | null): void {
-  const prefix = `${orgKeyPart(orgId)}::`;
-  for (const key of cache.keys()) {
-    if (key.startsWith(prefix)) cache.delete(key);
-  }
-}
-
-/** Reset entire cache. For testing only. */
-export function _resetPatternCache(): void {
-  cache.clear();
-}
-
-// ---------------------------------------------------------------------------
-// Keyword extraction & relevance scoring
-// ---------------------------------------------------------------------------
-
-/** Common SQL/English stop words to ignore during keyword matching. */
-const STOP_WORDS = new Set([
-  "the", "a", "an", "is", "are", "was", "were", "be", "been", "being",
-  "have", "has", "had", "do", "does", "did", "will", "would", "shall",
-  "should", "may", "might", "must", "can", "could", "and", "but", "or",
-  "nor", "not", "so", "yet", "for", "of", "in", "on", "at", "to", "by",
-  "with", "from", "as", "into", "about", "between", "through", "after",
-  "before", "above", "below", "up", "down", "out", "off", "over", "under",
-  "then", "than", "that", "this", "these", "those", "what", "which", "who",
-  "whom", "how", "when", "where", "why", "all", "each", "every", "both",
-  "few", "more", "most", "other", "some", "such", "no", "only", "same",
-  "it", "its", "i", "me", "my", "we", "us", "our", "you", "your", "he",
-  "him", "his", "she", "her", "they", "them", "their",
-  // SQL keywords
-  "select", "from", "where", "join", "left", "right", "inner", "outer",
-  "on", "group", "order", "limit", "offset", "having", "union", "case",
-  "when", "else", "end", "as", "count", "sum", "avg", "min", "max",
-  "distinct", "null", "true", "false", "asc", "desc", "between", "like",
-  "insert", "update", "delete", "create", "alter", "drop", "table",
-  "index", "values", "set", "into", "not", "exists", "if", "and", "or",
-]);
 
 // ---------------------------------------------------------------------------
 // Retrieval-query assembly
@@ -210,49 +111,12 @@ export function buildRetrievalQuery(
   return collected.reverse().join(" ").trim();
 }
 
-/** Extract meaningful keywords from a text string. */
-export function extractKeywords(text: string): Set<string> {
-  const words = text
-    .toLowerCase()
-    .replace(/[^a-z0-9_]/g, " ")
-    .split(/\s+/)
-    .filter((w) => w.length > 1 && !STOP_WORDS.has(w));
-  return new Set(words);
-}
-
-/** Score a pattern's relevance to a set of question keywords. */
-function scorePattern(
-  pattern: ApprovedPatternRow,
-  questionKeywords: Set<string>,
-): number {
-  const patternText = [
-    pattern.pattern_sql,
-    pattern.description ?? "",
-    pattern.source_entity ?? "",
-  ].join(" ");
-
-  const patternKeywords = extractKeywords(patternText);
-  let overlap = 0;
-  for (const kw of questionKeywords) {
-    if (patternKeywords.has(kw)) overlap++;
-  }
-  return overlap;
-}
-
 // ---------------------------------------------------------------------------
-// Public API
+// Retrieval API
 // ---------------------------------------------------------------------------
 
 /** Default maximum patterns to inject into the system prompt. */
 const DEFAULT_MAX_PATTERNS = 10;
-
-/**
- * Floor for the perf weight. A pattern arbitrarily slower than the budget is
- * down-weighted to at most this factor — never to zero — so slow-but-relevant
- * patterns stay PRESENT in retrieval as reference, just ranked below faster
- * peers. Keeping the floor > 0 is what makes this a down-weight, not a filter.
- */
-const MIN_PERF_WEIGHT = 0.5;
 
 export interface RelevantPattern {
   sourceEntity: string | null;
@@ -261,66 +125,6 @@ export interface RelevantPattern {
   /** Rolling-mean wall-clock (ms) of this pattern's runs, or null until first
    *  observed. Surfaced in the injected context so the agent can judge cost. */
   avgDurationMs: number | null;
-}
-
-/**
- * Latency down-weight factor in [{@link MIN_PERF_WEIGHT}, 1].
- *
- * - Unknown latency (null / non-finite / negative) → 1.0: we don't penalize a
- *   pattern whose speed we've never measured.
- * - At or under budget → 1.0: no penalty.
- * - Beyond budget → decays toward the floor as `budget / avg` shrinks, so a
- *   pattern 2× the budget keeps ~0.75 of its weight and one 10× keeps ~0.55.
- *
- * Pure helper, exported for direct testing.
- */
-export function perfWeight(avgDurationMs: number | null, budgetMs: number): number {
-  if (avgDurationMs === null || !Number.isFinite(avgDurationMs) || avgDurationMs < 0) return 1;
-  if (!Number.isFinite(budgetMs) || budgetMs <= 0) return 1;
-  if (avgDurationMs <= budgetMs) return 1;
-  const ratio = budgetMs / avgDurationMs; // in (0, 1)
-  return MIN_PERF_WEIGHT + (1 - MIN_PERF_WEIGHT) * ratio;
-}
-
-/** One scored pattern after keyword + perf weighting. Exported for testing. */
-export interface ScoredPattern {
-  pattern: ApprovedPatternRow;
-  keywordScore: number;
-  perfWeight: number;
-  score: number;
-}
-
-/**
- * Rank approved patterns for a question, down-weighting slow ones (PRD #3617
- * B-2). PURE: no DB, no settings — takes the candidate rows, the question
- * keywords, and the latency budget.
- *
- * The eligibility filter stays on RAW keyword overlap (`keywordScore > 0`), so
- * latency never DROPS a relevant pattern — it only reorders. The combined score
- * is `keywordScore × perfWeight`, which lets a much-more-relevant slow pattern
- * still outrank a barely-relevant fast one (relevance dominates), while a fast
- * pattern wins among similarly-relevant peers. Ties break on confidence, then
- * on lower latency (unknown latency sorts last).
- */
-export function rankPatterns(
-  patterns: readonly ApprovedPatternRow[],
-  questionKeywords: Set<string>,
-  opts: { latencyBudgetMs: number; maxPatterns: number },
-): ScoredPattern[] {
-  return patterns
-    .map((pattern): ScoredPattern => {
-      const keywordScore = scorePattern(pattern, questionKeywords);
-      const weight = perfWeight(pattern.avg_duration_ms, opts.latencyBudgetMs);
-      return { pattern, keywordScore, perfWeight: weight, score: keywordScore * weight };
-    })
-    .filter((s) => s.keywordScore > 0)
-    .toSorted(
-      (a, b) =>
-        b.score - a.score ||
-        b.pattern.confidence - a.pattern.confidence ||
-        (a.pattern.avg_duration_ms ?? Infinity) - (b.pattern.avg_duration_ms ?? Infinity),
-    )
-    .slice(0, opts.maxPatterns);
 }
 
 /**
